refactor: extract common permission logic
