+++ packages/nestjs-authorization/package.json
@@ -35,6 +35,7 @@
     "build": "tsdx build",
     "test": "jest",
     "test:watch": "jest --watch",
+    "test:cov": "jest --coverage",
     "lint": "tsdx lint",
     "semantic-release": "semantic-release",
     "upgrade-interactive": "npx npm-check --update"
