+++ packages/nestjs-authorization/src/authorization.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
+import { PermissionService } from './permission.service';
 import { ConditionService } from './condition.service';
 
 @Module({
   imports: [],
-  providers: [ConditionService],
-  exports: [ConditionService],
+  providers: [ConditionService, PermissionService],
+  exports: [ConditionService, PermissionService],
 })
 export class AuthorizationModule {}
