+++ packages/nestjs-authorization/src/permission.guard.spec.ts
@@ -9,6 +9,7 @@ import { ConditionService } from './condition.service';
 import { AbstractPermissionProvider } from './permission.provider';
 import { SimplePermissionSet } from './permission-set';
 import { Action, CreateAction } from './action';
+import { PermissionService } from './permission.service';
 
 @Injectable()
 class AlwaysTrueCondition extends AbstractCondition {
@@ -90,6 +91,7 @@ function createTestModule(actions: Action[]) {
     imports: [],
     providers: [
       ConditionService,
+      PermissionService,
       AlwaysFalseCondition,
       AlwaysTrueCondition,
       {
@@ -107,7 +109,7 @@ function createTestModule(actions: Action[]) {
   }).compile();
 }
 
-describe('AuthActionGuard', () => {
+describe('PermissionGuard', () => {
   it('canActivate should return false if user has NOT permission', async () => {
     const app = await initApp(['write']);
     const guard = await app.get('GUARD');
