+++ packages/nestjs-authorization/src/permission.guard.ts
@@ -5,10 +5,9 @@ import {
   Injectable,
   mixin,
 } from '@nestjs/common';
-import { isFunctionalCondition } from './condition';
 import { Action } from './action';
 import { PermissionProvider } from './permission.provider';
-import { ConditionService } from './condition.service';
+import { PermissionService } from './permission.service';
 
 export const PermissionGuard = (actions: Action[]) => {
   @Injectable()
@@ -16,50 +15,13 @@ export const PermissionGuard = (actions: Action[]) => {
     constructor(
       @Inject('PERMISSION_PROVIDER')
       private permissionProvider: PermissionProvider,
-      private conditionService: ConditionService
+      private permissionService: PermissionService
     ) {}
 
     async canActivate(context: ExecutionContext): Promise<boolean> {
       const req = context.switchToHttp().getRequest();
       const permissionSet = await this.permissionProvider.getPermissionSet(req);
-      const result = actions.reduce(async (isAllowedPromise, action) => {
-        const isAllowed = await isAllowedPromise;
-        if (isAllowed) {
-          return Promise.resolve(true);
-        }
-
-        if (typeof action === 'string') {
-          return await permissionSet.isAllowed(action, context);
-        }
-
-        if (action.condition === undefined) {
-          return await permissionSet.isAllowed(action.name, context);
-        }
-
-        // conditional action by function
-        if (isFunctionalCondition(action.condition)) {
-          return action.condition(context)
-            ? await permissionSet.isAllowed(action.name, context)
-            : false;
-        }
-
-        // conditional action by type
-        const condition = await this.conditionService.findByType(
-          action.condition
-        );
-        if (
-          condition &&
-          (await condition.check(context, {
-            action: action.name,
-            permission: undefined,
-          }))
-        ) {
-          return await permissionSet.isAllowed(action.name, context);
-        }
-
-        return false;
-      }, Promise.resolve(false));
-      return await result;
+      return this.permissionService.areAllowed(actions, permissionSet, context);
     }
   }
